refactor: address CR comments
affects: django-testbed
